Redefine Model.this with `enable_default_scope`
Since `enable_default_scope` changes the dataset for a model, every
instance of that model is affected in its `this` method (which is used
for e.g. updating). This means that once (soft-)deleted, an instance can
no longer be updated, because its `this` method returns something like:

SELECT * FROM model WHERE ((`id` = 1) AND (`deleted_at` IS NULL));

This returns 0 rows instead of 1, and any changes therefore fail. This
includes the `recover` method.