Fix inappropriate true value for `#deleted?` on new models
Fixes an issue where `#deleted?` would return true on new (unsaved)
models when the `deleted_column_default` value is non-nil. Since this
column is always `nil` until the model is persisted, the previous
definition of `#deleted?` would compare this `nil` value with the
non-nil value in `deleted_column_default` and so always return true.